requirements.txt
git+https://github.com/mmabrouk/chatgpt-wrapper
pyyaml==6.0
marshmallow_dataclass
loguru==0.6.0
redis
tweepy==4.12.1